requirements.txt
asgiref==3.6.0
Django==3.2
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
PyMySQL==1.0.2
pytz==2022.7
sqlparse==0.4.3
